added a sorting function for the String list read. Renamed the package and the class to keep consistency (counter to processor)
